Test: 오류나는 테스트코드 수정
Resolves: #238
