+++ src/main/java/com/explorer/gabom/global/security/userdetails/CustomUserDetails.java
@@ -61,8 +61,4 @@ public class CustomUserDetails implements UserDetails {
 	public User getUser() {
 		return this.user;
 	}
-
-	public static CustomUserDetails fromUser(User user) {
-		return new CustomUserDetails(user.getId(), user.getEmail(), user.getPassword(), user.getUserRole(), user);
-	}
 }
